backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
chromadb>=0.4.22
sentence-transformers>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.0.0
